requirements.txt
dj-database-url==0.5.0
Django==2.2.1
django-filter==2.1.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.9.4
get==2019.4.13
gunicorn==19.9.0
Pillow==6.0.0
post==2019.4.13
psycopg2==2.8.2
public==2019.4.13
pytz==2019.1
query-string==2019.4.13
request==2019.4.13
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.2
